+++ pineappl_cli/src/convolve.rs
@@ -37,7 +37,13 @@ pub struct Opts {
         value_delimiter = ',',
         value_parser = helpers::parse_order
     )]
-    orders: Vec<(u8, u8)>,
+    orders: Vec<(u32, u32)>,
+    /// Set the variation of the renormalization scale.
+    #[arg(default_value = "1.0", long, num_args = 1)]
+    xir: f64,
+    /// Set the variation of the factorization scale.
+    #[arg(default_value = "1.0", long, num_args = 1)]
+    xif: f64,
     /// Set the number of fractional digits shown for absolute numbers.
     #[arg(default_value_t = 7, long, value_name = "ABS")]
     digits_abs: usize,
@@ -52,13 +58,13 @@ impl Subcommand for Opts {
         let mut conv_funs_0 = helpers::create_conv_funs(&self.conv_funs[0])?;
         let bins: Vec<_> = self.bins.iter().cloned().flatten().collect();
 
-        let results = helpers::convolve(
+        let results = helpers::convolve_scales(
             &grid,
             &mut conv_funs_0,
             &self.orders,
             &bins,
             &[],
-            1,
+            &[(self.xir, self.xif)],
             if self.integrated {
                 ConvoluteMode::Integrated
             } else {